Did a few of the easier review suggestions. - Still workin on the rest...
