+++ loaforcsSoundAPI.Patcher/loaforcsSoundAPI.Patcher.csproj
@@ -3,8 +3,6 @@
     <PropertyGroup>
         <AssemblyName>me.loaforc.soundapi.patcher</AssemblyName>
         <Product>loaforcsSoundAPI.Patcher</Product>
-        <!-- Change to whatever version you're currently on. -->
-        <Version>1.0.0</Version>
     </PropertyGroup>
 
     <!-- Project Properties -->
@@ -13,14 +11,6 @@
         <RootNamespace>loaforcsSoundAPI.Patcher</RootNamespace>
         <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
         <LangVersion>preview</LangVersion>
-    </PropertyGroup>
-
-    <!-- Set NuGet Sources -->
-    <PropertyGroup>
-        <RestoreAdditionalProjectSources>
-            https://api.nuget.org/v3/index.json;
-            https://nuget.bepinex.dev/v3/index.json
-        </RestoreAdditionalProjectSources>
     </PropertyGroup>
 
     <!-- Embed Debug Symbols for Easier Debugging -->
