+++ loaforcsSoundAPI/SoundPacks/Data/Conditions/DefaultConditionContext.cs
@@ -2,8 +2,8 @@
 
 namespace loaforcsSoundAPI.SoundPacks.Data.Conditions;
 
-struct DefaultConditionContext(AudioSource source) : IContext {
+class DefaultConditionContext(AudioSource source) : IContext {
 	internal static readonly DefaultConditionContext DEFAULT = new DefaultConditionContext(null);
 
-	public readonly AudioSource Source => source;
+	public AudioSource Source => source;
 }
